Move deployment to its own module
